requirements.txt
altgraph==0.17.3
anyio==3.7.0
async-timeout==4.0.2
beautifulsoup4==4.12.2
Brotli==1.0.9
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
darkdetect==0.7.1
decorator==5.1.1
Deprecated==1.2.14
exceptiongroup==1.1.1
fastapi==0.95.2
ffmpy==0.3.0
filelock==3.12.0
h11==0.14.0
idna==3.4
jaconv==0.3.4
markdown-it-py==2.2.0
materialyoucolor==1.2.0
mdurl==0.1.2
mutagen==1.46.0
numpy==1.24.3
Pillow==9.5.0
platformdirs==3.5.1
pycryptodomex==3.18.0
pydantic==1.10.8
Pygments==2.15.1
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
pykakasi==2.2.1
pyqtdarktheme==2.1.0
PySide6==6.2.4
python-dotenv==1.0.0
python-slugify==8.0.1
pytube==12.1.3
rapidfuzz==2.15.1
redis==4.5.5
requests==2.31.0
requests-file==1.5.1
rich==13.3.5
savify==2.4.0
scipy==1.10.1
shiboken6==6.2.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
spotdl==4.1.11
spotipy==2.23.0
starlette==0.27.0
syncedlyrics==0.5.0
text-unidecode==1.3
tldextract==3.4.4
typing-extensions==4.6.2
Unidecode==1.3.6
urllib3==2.0.2
uvicorn==0.22.0
validators==0.20.0
websockets==11.0.3
wrapt==1.15.0
yt-dlp==2023.3.4
ytmusicapi==0.24.1
